server/python/requirements.txt
openai-whisper
transformers
torch
